Fix errors that occur when Cassandra has NULL data
While trying to use and develop with this Demo, I ran into an issue with Dataflow where it errored out with NullPointerExceptions. 

Turns out some data from Cassandra was NULL. This change *should* solve this problem. This is because you would now be accepting an Object in the mutation and converting it to a string. If the string is Null you can set it to NULL which seems to properly work on BT. 

Please check and confirm or loop back in with me for questions. There might be more unintended consequences of this that I'm not aware of.